root part 5 - login logout handler
