+++ webapp/channels/src/components/root/actions.ts
@@ -15,6 +15,8 @@ import type {ThunkActionFunc} from 'mattermost-redux/types/actions';
 
 import * as GlobalActions from 'actions/global_actions';
 
+import {StoragePrefixes} from 'utils/constants';
+
 export function redirectToOnboardingOrDefaultTeam(history: History): ThunkActionFunc<void> {
     return async (dispatch, getState) => {
         const state = getState();
@@ -58,5 +60,32 @@ export function redirectToOnboardingOrDefaultTeam(history: History): ThunkAction
         }
 
         GlobalActions.redirectUserToDefaultTeam();
+    };
+}
+
+export function handleLoginLogoutSignal(e: StorageEvent): ThunkActionFunc<void> {
+    return (dispatch, getState) => {
+    // when one tab on a browser logs out, it sets __logout__ in localStorage to trigger other tabs to log out
+        const isNewLocalStorageEvent = (event: StorageEvent) => event.storageArea === localStorage && event.newValue;
+
+        if (e.key === StoragePrefixes.LOGOUT && isNewLocalStorageEvent(e)) {
+            console.log('detected logout from a different tab'); //eslint-disable-line no-console
+            GlobalActions.emitUserLoggedOutEvent('/', false, false);
+        }
+        if (e.key === StoragePrefixes.LOGIN && isNewLocalStorageEvent(e)) {
+            const isLoggedIn = getCurrentUser(getState());
+
+            // make sure this is not the same tab which sent login signal
+            // because another tabs will also send login signal after reloading
+            if (isLoggedIn) {
+                return;
+            }
+
+            // detected login from a different tab
+            function reloadOnFocus() {
+                location.reload();
+            }
+            window.addEventListener('focus', reloadOnFocus);
+        }
     };
 }
